pyproject.toml
[tool.poetry]
name = "repl-nix-redlinemaster"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.44.0"
flask = "^3.0.3"
pypdf2 = "^3.0.1"
diff-match-patch = "^20230430"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
